(#65) Organize views in views-folder
