+++ d2qc/d2qc/data/views/cast.py
@@ -1,0 +1,10 @@
+from rest_framework import viewsets
+
+from d2qc.data.models import Cast
+from d2qc.data.serializers import CastSerializer
+
+
+class CastViewSet(viewsets.ModelViewSet):
+
+    queryset = Cast.objects.all()
+    serializer_class = CastSerializer
