+++ d2qc/d2qc/data/views/data_unit.py
@@ -1,0 +1,10 @@
+from rest_framework import viewsets
+
+from d2qc.data.models import DataUnit
+from d2qc.data.serializers import DataUnitSerializer
+
+
+class DataUnitViewSet(viewsets.ModelViewSet):
+
+    queryset = DataUnit.objects.all().order_by('-created')
+    serializer_class = DataUnitSerializer
