+++ d2qc/d2qc/data/views/data_value.py
@@ -1,0 +1,10 @@
+from rest_framework import viewsets
+
+from d2qc.data.models import DataValue
+from d2qc.data.serializers import DataValueSerializer
+
+
+class DataValueViewSet(viewsets.ModelViewSet):
+
+    queryset = DataValue.objects.all().order_by('-created')
+    serializer_class = DataValueSerializer
