+++ d2qc/d2qc/data/views/depth.py
@@ -1,0 +1,10 @@
+from rest_framework import viewsets
+
+from d2qc.data.models import Depth
+from d2qc.data.serializers import DepthSerializer
+
+
+class DepthViewSet(viewsets.ModelViewSet):
+
+    queryset = Depth.objects.all()
+    serializer_class = DepthSerializer
