+++ d2qc/d2qc/data/views/index_page.py
@@ -1,0 +1,4 @@
+from django.views.generic import TemplateView
+
+class IndexPage(TemplateView):
+    template_name = 'index.html'
